+++ packages/ra-ui-materialui/src/theme/index.ts
@@ -1,0 +1,2 @@
+export * from './defaultTheme';
+export * from './nanoTheme';
